rock-paper-scissors-shared-main/requirements.txt
absl-py==0.14.0
aioice==0.7.6
aiortc==1.3.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.2.0
av==8.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.2.0
entrypoints==0.3
emoji==2.1.0
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.26
google-crc32c==1.3.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.8.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.8
numpy==1.19.5
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyarrow==6.0.1
pycodestyle==2.7.0
pycparser==2.21
pydeck==0.7.1
pydub==0.25.1
pyee==9.0.3
Pygments==2.10.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
PySide2==5.15.2
PySimpleGUI==4.50.0
PySimpleGUIQt==0.35.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
shiboken2==5.15.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
streamlit==1.5.0
streamlit-webrtc==0.34.2
terminado==0.13.1
testpath==0.5.0
threadpoolctl==2.2.0
tk==0.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
